move api options into cli
